+++ src/main/java/name/divinityunbound/datagen/ModModelProvider.java
@@ -150,6 +150,11 @@ public class ModModelProvider extends FabricModelProvider {
         itemModelGenerator.registerArmor((ArmorItem) ModItems.SPACE_FORGED_LEGGINGS);
         itemModelGenerator.registerArmor((ArmorItem) ModItems.SPACE_FORGED_BOOTS);
 
+        itemModelGenerator.registerArmor((ArmorItem) ModItems.EXPERIENCE_HELMET);
+        itemModelGenerator.registerArmor((ArmorItem) ModItems.EXPERIENCE_CHESTPLATE);
+        itemModelGenerator.registerArmor((ArmorItem) ModItems.EXPERIENCE_LEGGINGS);
+        itemModelGenerator.registerArmor((ArmorItem) ModItems.EXPERIENCE_BOOTS);
+
         /* Foci */
         itemModelGenerator.register(ModItems.CELESTITE_COAL_FOCUS, Models.GENERATED);
         itemModelGenerator.register(ModItems.CELESTITE_IRON_FOCUS, Models.GENERATED);